Migrate GraphQL dotnet to HotChocolate

+++ Source/Coco.Server/Api.Identity/Resolvers/UserResolver.cs
@@ -9,6 +9,10 @@ using Coco.Entities.Dtos.User;
 using System.Collections.Generic;
 using Api.Identity.Resolvers.Contracts;
 using HotChocolate.Resolvers;
+using Coco.Common.Exceptions;
+using Coco.Commons.Models;
+using Coco.Common.Const;
+using Api.Identity.Models;
 
 namespace Api.Identity.Resolvers
 {
@@ -23,7 +27,7 @@ namespace Api.Identity.Resolvers
             _loginManager = loginManager;
         }
 
-        public async Task<ApiResult> UpdateUserInfoItemAsync(IResolverContext context)
+        public async Task<UpdatePerItemModel> UpdateUserInfoItemAsync(IResolverContext context)
         {
             try
             {
@@ -44,25 +48,20 @@ namespace Api.Identity.Resolvers
 
                 var result = await _userManager.UpdateInfoItemAsync(model, currentUser.UserIdentityId, currentUser.AuthenticationToken);
 
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
-
                 return result;
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> SignoutAsync(IDictionary<string, object> userContext)
+        public async Task<IApiResult> SignoutAsync(IResolverContext context)
         {
             try
             {
-                var sessionContext = userContext["SessionContext"] as ISessionContext;
-
+                var sessionContext = context.ContextData["SessionContext"] as ISessionContext;
                 if (sessionContext == null || sessionContext.CurrentUser == null)
                 {
                     throw new UnauthorizedAccessException();
@@ -70,17 +69,26 @@ namespace Api.Identity.Resolvers
 
                 var currentUser = sessionContext.CurrentUser;
 
-                var result = await _loginManager.LogoutAsync(currentUser.UserIdentityId, currentUser.AuthenticationToken);
+                var isLoggedout = await _loginManager.LogoutAsync(currentUser.UserIdentityId, currentUser.AuthenticationToken);
 
-                return result;
+                if (!isLoggedout)
+                {
+                    return ApiResult.Failed(new CommonError()
+                    {
+                        Code = ErrorMessageConst.EXCEPTION,
+                        Message = ErrorMessageConst.UN_EXPECTED_EXCEPTION
+                    });
+                }
+                return ApiResult.Success();
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> UpdateAvatarAsync(IResolverContext context)
+        public async Task<IApiResult> UpdateAvatarAsync(IResolverContext context)
         {
             try
             {
@@ -88,22 +96,18 @@ namespace Api.Identity.Resolvers
                 var userContext = context.ContextData["SessionContext"] as ISessionContext;
 
                 model.UserPhotoType = UserPhotoTypeEnum.Avatar;
-                var result = await _userManager.UpdateAvatarAsync(model, userContext.CurrentUser.Id);
+                var result = await _userManager.UpdateAvatarAsync(model, userContext.CurrentUser);
 
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
-
-                return result;
+                return ApiResult.Success(result);
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> UpdateCoverAsync(IResolverContext context)
+        public async Task<IApiResult> UpdateCoverAsync(IResolverContext context)
         {
             try
             {
@@ -111,91 +115,94 @@ namespace Api.Identity.Resolvers
                 var userContext = context.ContextData["SessionContext"] as ISessionContext;
 
                 model.UserPhotoType = UserPhotoTypeEnum.Cover;
-                var result = await _userManager.UpdateCoverAsync(model, userContext.CurrentUser.Id);
-                
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
+                var result = await _userManager.UpdateCoverAsync(model, userContext.CurrentUser);
 
-                return result;
+                return ApiResult.Success(result);
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> DeleteAvatarAsync(IResolverContext context)
+        public async Task<IApiResult> DeleteAvatarAsync(IResolverContext context)
         {
             try
             {
-                var model = GenerateUserPhotoModel(context);
+                var criterias = context.Argument<PhotoDeleteModel>("criterias");
                 var userContext = context.ContextData["SessionContext"] as ISessionContext;
+
+                if (!criterias.CanEdit)
+                {
+                    throw new UnauthorizedAccessException();
+                }
+
+                if (userContext == null || userContext.CurrentUser == null)
+                {
+                    throw new UnauthorizedAccessException();
+                }
 
                 var result = await _userManager.DeleteUserPhotoAsync(userContext.CurrentUser.Id, UserPhotoTypeEnum.Avatar);
-
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
-
-                return result;
+                return ApiResult.Success(result);
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> DeleteCoverAsync(IResolverContext context)
+        public async Task<IApiResult> DeleteCoverAsync(IResolverContext context)
         {
             try
             {
-                var model = GenerateUserPhotoModel(context);
+                var criterias = context.Argument<PhotoDeleteModel>("criterias");
                 var userContext = context.ContextData["SessionContext"] as ISessionContext;
+
+                if (!criterias.CanEdit)
+                {
+                    throw new UnauthorizedAccessException();
+                }
+
+                if (userContext == null || userContext.CurrentUser == null)
+                {
+                    throw new UnauthorizedAccessException();
+                }
 
                 var result = await _userManager.DeleteUserPhotoAsync(userContext.CurrentUser.Id, UserPhotoTypeEnum.Cover);
 
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
-
-                return result;
+                return ApiResult.Success(result);
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> UpdateIdentifierAsync(IResolverContext context)
+        public async Task<UserIdentifierUpdateDto> UpdateIdentifierAsync(IResolverContext context)
         {
             try
             {
-                var user = context.Argument<ApplicationUser>("user");
+                var criterias = context.Argument<UserIdentifierUpdateDto>("criterias");
                 var userContext = context.ContextData["SessionContext"] as ISessionContext;
 
                 var currentUser = userContext.CurrentUser;
-                user.Id = currentUser.Id;
+                currentUser.Lastname = criterias.Lastname;
+                currentUser.Firstname = criterias.Firstname;
+                currentUser.DisplayName = criterias.DisplayName;
 
-                var result = await _userManager.UpdateIdentifierAsync(user, currentUser.UserIdentityId, currentUser.AuthenticationToken);
-
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
-
-                return result;
+                return await _userManager.UpdateIdentifierAsync(currentUser, currentUser.UserIdentityId, currentUser.AuthenticationToken);
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return null;
             }
         }
 
-        public async Task<ApiResult> UpdatePasswordAsync(IResolverContext context)
+        public async Task<UserTokenResult> UpdatePasswordAsync(IResolverContext context)
         {
             try
             {
@@ -212,23 +219,19 @@ namespace Api.Identity.Resolvers
 
                 var result = await _userManager.ChangePasswordAsync(currentUser.Id, model.CurrentPassword, model.NewPassword);
 
-                if (!result.IsSucceed)
-                {
-                    HandleContextError(context, result.Errors);
-                }
-
                 return result;
             }
             catch (Exception ex)
             {
-                throw;
+                HandleContextError(context, ex);
+                return new UserTokenResult(false);
             }
         }
 
         #region Privates
-        private UpdateUserPhotoDto GenerateUserPhotoModel(IResolverContext context)
+        private UserPhotoUpdateDto GenerateUserPhotoModel(IResolverContext context)
         {
-            var model = context.Argument<UpdateUserPhotoDto>("criterias");
+            var model = context.Argument<UserPhotoUpdateDto>("criterias");
             var userContext = context.ContextData["SessionContext"] as ISessionContext;
 
             if (!model.CanEdit)
